FileCounter: Remove and replace test with a ScannerIntegrationFunTest
This allows for an integration test for the scanner, but removes
the superfluous file counter scanner, now that fileCount is removed.
