+++ scanner/src/funTest/assets/dummy-expected-output-for-analyzer-result.yml
@@ -214,7 +214,7 @@ scanner:
       Gradle:org.ossreviewtoolkit.gradle.example:lib:1.0.0:
       - provenance: {}
         scanner:
-          name: "FileCounter"
+          name: "Dummy"
           version: "1.0"
           configuration: ""
         summary:
@@ -225,7 +225,7 @@ scanner:
           copyrights: []
           issues:
           - timestamp: "1970-01-01T00:00:00Z"
-            source: "FileCounter"
+            source: "Dummy"
             message: "Could not download 'Gradle:org.ossreviewtoolkit.gradle.example:lib:1.0.0':\
               \ DownloadException: Download failed for 'Gradle:org.ossreviewtoolkit.gradle.example:lib:1.0.0'.\n\
               Suppressed: DownloadException: No VCS URL provided for 'Gradle:org.ossreviewtoolkit.gradle.example:lib:1.0.0'.\
@@ -241,7 +241,7 @@ scanner:
               value: "a6c32b40bf3d76eca54e3c601e5d1470c86fcdfa"
               algorithm: "SHA-1"
         scanner:
-          name: "FileCounter"
+          name: "Dummy"
           version: "1.0"
           configuration: ""
         summary:
@@ -258,7 +258,7 @@ scanner:
               value: "f7d878153e86a1cdddf6b37850e00a9f8bff726f"
               algorithm: "SHA-1"
         scanner:
-          name: "FileCounter"
+          name: "Dummy"
           version: "1.0"
           configuration: ""
         summary:
@@ -275,7 +275,7 @@ scanner:
               value: "f0770f7f0472bf120ada47beecadce4056fbd20a"
               algorithm: "SHA-1"
         scanner:
-          name: "FileCounter"
+          name: "Dummy"
           version: "1.0"
           configuration: ""
         summary:
@@ -292,7 +292,7 @@ scanner:
               value: "1dc37250fbc78e23a65a67fbbaf71d2e9cbc3c0b"
               algorithm: "SHA-1"
         scanner:
-          name: "FileCounter"
+          name: "Dummy"
           version: "1.0"
           configuration: ""
         summary:
