+++ scanner/src/main/resources/META-INF/services/org.ossreviewtoolkit.scanner.ScannerFactory
@@ -1,6 +1,5 @@
 org.ossreviewtoolkit.scanner.scanners.Askalono$Factory
 org.ossreviewtoolkit.scanner.scanners.BoyterLc$Factory
-org.ossreviewtoolkit.scanner.scanners.FileCounter$Factory
 org.ossreviewtoolkit.scanner.scanners.Licensee$Factory
 org.ossreviewtoolkit.scanner.scanners.fossid.FossId$Factory
 org.ossreviewtoolkit.scanner.scanners.scancode.ScanCode$Factory
